pyproject.toml
[project]
name = "ocr"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.25.2",
    "docling>=2.74.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
